Move create_data_frame to sim utils
